requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
dataclasses-json==0.6.7
datasets==4.7.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
filelock==3.25.1
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.0
google-genai==1.66.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
idna==3.11
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==1.2.10
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
ragas==0.4.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
scikit-network==0.33.5
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
